Calculate start and finish for relays
We need to know when a team has started in order to activate its virtual time, before it downloads and after its start time. That information is in the list of runners.

We need the team’s finish_time so we can know whether the team has finished. In a relay-less OPIE, the finish_time indicates who finishes; if finish_time is null, it means they haven’t finished.